Refactor Edit, List and View into separate files
