+++ Development/src/components/CardFormIterator.js
@@ -7,7 +7,7 @@ import React, {
 import PropTypes from 'prop-types';
 import get from 'lodash/get';
 import { Card, CardContent, Grid } from '@material-ui/core';
-import { FormInput } from 'ra-ui-materialui';
+import { FormInput } from 'react-admin';
 // Derived from react-admin component
 export class CardFormIterator extends Component {
     constructor(props) {
