+++ Development/src/components/QueryVersion.js
@@ -1,0 +1,9 @@
+import Cookies from 'universal-cookie';
+const cookies = new Cookies();
+
+const QueryVersion = () => {
+    let url = cookies.get('Query API');
+    return url.match(/([^/]+)(?=\/?$)/g)[0];
+};
+
+export default QueryVersion;
