+++ Development/src/pages/devices/DevicesList.js
@@ -1,0 +1,126 @@
+import React, { useState } from 'react';
+import {
+    Card,
+    CardContent,
+    Table,
+    TableBody,
+    TableCell,
+    TableHead,
+    TableRow,
+} from '@material-ui/core';
+import { Button, ShowButton, Title } from 'react-admin';
+import FilterField from '../../components/FilterField';
+import JsonIcon from '../../components/JsonIcon';
+import PaginationButton from '../../components/PaginationButton';
+import dataProvider from '../../dataProvider';
+
+const DevicesList = () => {
+    const firstLoad = async () => {
+        const params = {
+            filter: {},
+        };
+        const dataObject = await dataProvider('GET_LIST', 'devices', params);
+        setData(dataObject);
+    };
+
+    const [data, setData] = useState(firstLoad);
+
+    const nextPage = async label => {
+        const dataObject = await dataProvider(label, 'devices');
+        setData(dataObject);
+    };
+
+    const [filterState, setFilterState] = useState({});
+
+    const changeFilter = async (filterValue, name) => {
+        let filter = filterState;
+        if (filterValue) {
+            filter[name] = filterValue;
+        } else {
+            delete filter[name];
+        }
+        const filteredDataObject = await dataProvider('GET_LIST', 'devices', {
+            filter: filter,
+        });
+        setFilterState(filter);
+        setData(filteredDataObject);
+    };
+
+    const clearFilter = async () => {
+        setData(firstLoad());
+        setFilterState({});
+    };
+
+    if (data.hasOwnProperty('data')) {
+        return (
+            <Card>
+                <Title title={'Devices'} />
+                <CardContent>
+                    <Button
+                        label={'Raw'}
+                        href={data.url}
+                        style={{ float: 'right' }}
+                        title={'View raw'}
+                    >
+                        <JsonIcon />
+                    </Button>
+                    <Table>
+                        <TableHead>
+                            <TableRow>
+                                <TableCell
+                                    style={{
+                                        minWidth: '240px',
+                                        paddingLeft: '32px',
+                                    }}
+                                >
+                                    Label{' '}
+                                    <FilterField
+                                        name="label"
+                                        setFilter={changeFilter}
+                                    />
+                                </TableCell>
+                                <TableCell style={{ minWidth: '240px' }}>
+                                    Type{' '}
+                                    <FilterField
+                                        name="type"
+                                        setFilter={changeFilter}
+                                    />
+                                </TableCell>
+                            </TableRow>
+                        </TableHead>
+                        <TableBody>
+                            {data.data.map(item => (
+                                <TableRow key={item.id}>
+                                    <TableCell component="th" scope="row">
+                                        <ShowButton
+                                            style={{
+                                                textTransform: 'none',
+                                            }}
+                                            basePath="/devices"
+                                            record={item}
+                                            label={item.label}
+                                        />
+                                    </TableCell>
+                                    <TableCell>{item.type}</TableCell>
+                                </TableRow>
+                            ))}
+                        </TableBody>
+                    </Table>
+                    <br />
+                    <PaginationButton label="FIRST" nextPage={nextPage} />
+                    <PaginationButton label="PREV" nextPage={nextPage} />
+                    <PaginationButton label="NEXT" nextPage={nextPage} />
+                    <PaginationButton label="LAST" nextPage={nextPage} />
+                    <Button
+                        onClick={() => clearFilter()}
+                        label="Clear All Filters"
+                    />
+                </CardContent>
+            </Card>
+        );
+    } else {
+        return <div />;
+    }
+};
+
+export default DevicesList;
