+++ Development/src/pages/devices/index.js
@@ -1,0 +1,4 @@
+import DevicesList from './DevicesList';
+import DevicesShow from './DevicesShow';
+
+export { DevicesList, DevicesShow };
