+++ Development/src/pages/flows/index.js
@@ -1,0 +1,4 @@
+import FlowsList from './FlowsList';
+import FlowsShow from './FlowsShow';
+
+export { FlowsList, FlowsShow };
