+++ Development/src/pages/nodes/index.js
@@ -1,0 +1,4 @@
+import NodesList from './NodesList';
+import NodesShow from './NodesShow';
+
+export { NodesList, NodesShow };
