+++ Development/src/pages/queryapis/QueryAPIsList.js
@@ -1,31 +1,11 @@
 import React from 'react';
-import {
-    Datagrid,
-    List,
-    ListButton,
-    RichTextField,
-    Show,
-    ShowButton,
-    SimpleShowLayout,
-    TextField,
-} from 'react-admin';
-import ConnectButton from '../components/ConnectButton';
-import { CardActions, hr } from '@material-ui/core';
+import { CardActions } from '@material-ui/core';
+import { Datagrid, List, ShowButton, TextField } from 'react-admin';
+import ConnectButton from '../../components/ConnectButton';
 
 const QueryAPITitle = ({ record }) => {
     return <span>Query APIs{record ? `: ${record.name}` : ''}</span>;
 };
-
-const cardActionsStyle = {
-    zIndex: 2,
-    float: 'right',
-};
-
-const QueryAPIShowActions = ({ basePath }) => (
-    <CardActions title={<QueryAPITitle />} style={cardActionsStyle}>
-        <ListButton title={'Return to ' + basePath} basePath={basePath} />
-    </CardActions>
-);
 
 const QueryAPIPagination = () => {
     return (
@@ -66,7 +46,7 @@ const QueryAPIActions = ({
     </CardActions>
 );
 
-export const QueryAPIList = props => (
+const QueryAPIList = props => (
     <List
         title={<QueryAPITitle />}
         actions={<QueryAPIActions />}
@@ -90,24 +70,4 @@ export const QueryAPIList = props => (
     </List>
 );
 
-export const QueryAPIShow = props => (
-    <Show
-        title={<QueryAPITitle />}
-        actions={<QueryAPIShowActions />}
-        {...props}
-    >
-        <SimpleShowLayout>
-            <br />
-            <ConnectButton />
-            <TextField source="name" />
-            <hr />
-            <RichTextField source="addresses" />
-            <TextField label="Host Target" source="host_target" />
-            <TextField source="port" />
-            <hr />
-            <TextField label="API Protocol" source="txt.api_proto" />
-            <TextField label="API Versions" source="txt.api_ver" />
-            <TextField label="Priority" source="txt.pri" />
-        </SimpleShowLayout>
-    </Show>
-);
+export default QueryAPIList;
