+++ Development/src/pages/queryapis/index.js
@@ -1,0 +1,4 @@
+import QueryAPIList from './QueryAPIsList';
+import QueryAPIShow from './QueryAPIsShow';
+
+export { QueryAPIList, QueryAPIShow };
