+++ Development/src/pages/receivers/ReceiversList.js
@@ -1,0 +1,156 @@
+import React, { useState } from 'react';
+import {
+    Card,
+    CardContent,
+    Table,
+    TableBody,
+    TableCell,
+    TableHead,
+    TableRow,
+} from '@material-ui/core';
+import { Button, ShowButton, Title } from 'react-admin';
+import CheckIcon from '@material-ui/icons/Check';
+import ClearIcon from '@material-ui/icons/Clear';
+import Cookies from 'universal-cookie';
+import FilterField from '../../components/FilterField';
+import JsonIcon from '../../components/JsonIcon';
+import PaginationButton from '../../components/PaginationButton';
+import dataProvider from '../../dataProvider';
+
+const cookies = new Cookies();
+
+const QueryVersion = () => {
+    let url = cookies.get('Query API');
+    return url.match(/([^/]+)(?=\/?$)/g)[0];
+};
+
+const ReceiversList = () => {
+    const firstLoad = async () => {
+        const params = {
+            filter: {},
+        };
+        const dataObject = await dataProvider('GET_LIST', 'receivers', params);
+        setData(dataObject);
+    };
+
+    const [data, setData] = useState(firstLoad);
+
+    const nextPage = async label => {
+        const dataObject = await dataProvider(label, 'receivers');
+        setData(dataObject);
+    };
+
+    const [filterState, setFilterState] = useState({});
+
+    const changeFilter = async (filterValue, name) => {
+        let filter = filterState;
+        if (filterValue) {
+            filter[name] = filterValue;
+        } else {
+            delete filter[name];
+        }
+        const filteredDataObject = await dataProvider('GET_LIST', 'receivers', {
+            filter: filter,
+        });
+        setFilterState(filter);
+        setData(filteredDataObject);
+    };
+
+    const clearFilter = async () => {
+        setData(firstLoad());
+        setFilterState({});
+    };
+
+    if (data.hasOwnProperty('data')) {
+        return (
+            <Card>
+                <Title title={'Receivers'} />
+                <CardContent>
+                    <Button
+                        label={'Raw'}
+                        href={data.url}
+                        style={{ float: 'right' }}
+                        title={'View raw'}
+                    >
+                        <JsonIcon />
+                    </Button>
+                    <Table>
+                        <TableHead>
+                            <TableRow>
+                                <TableCell
+                                    style={{
+                                        minWidth: '240px',
+                                        paddingLeft: '32px',
+                                    }}
+                                >
+                                    Label{' '}
+                                    <FilterField
+                                        name="label"
+                                        setFilter={changeFilter}
+                                    />
+                                </TableCell>
+                                <TableCell style={{ minWidth: '245px' }}>
+                                    Format{' '}
+                                    <FilterField
+                                        name="format"
+                                        setFilter={changeFilter}
+                                    />
+                                </TableCell>
+                                <TableCell style={{ minWidth: '260px' }}>
+                                    Transport{' '}
+                                    <FilterField
+                                        name="transport"
+                                        setFilter={changeFilter}
+                                    />
+                                </TableCell>
+                                {QueryVersion() >= 'v1.2' && (
+                                    <TableCell>Active</TableCell>
+                                )}
+                            </TableRow>
+                        </TableHead>
+                        <TableBody>
+                            {data.data.map(item => (
+                                <TableRow key={item.id}>
+                                    <TableCell component="th" scope="row">
+                                        <ShowButton
+                                            style={{
+                                                textTransform: 'none',
+                                            }}
+                                            basePath="/receivers"
+                                            record={item}
+                                            label={item.label}
+                                        />
+                                    </TableCell>
+                                    <TableCell>{item.format}</TableCell>
+                                    <TableCell>{item.transport}</TableCell>
+                                    {QueryVersion() >= 'v1.2' && (
+                                        <TableCell>
+                                            {item.subscription.active ? (
+                                                <CheckIcon />
+                                            ) : (
+                                                <ClearIcon />
+                                            )}
+                                        </TableCell>
+                                    )}
+                                </TableRow>
+                            ))}
+                        </TableBody>
+                    </Table>
+                    <br />
+                    <PaginationButton label="FIRST" nextPage={nextPage} />
+                    <PaginationButton label="PREV" nextPage={nextPage} />
+                    <PaginationButton label="NEXT" nextPage={nextPage} />
+                    <PaginationButton label="LAST" nextPage={nextPage} />
+                    <Button
+                        onClick={() => clearFilter()}
+                        label="Clear All Filters"
+                    />
+                </CardContent>
+            </Card>
+        );
+    } else {
+        return <div />;
+    }
+};
+
+export default ReceiversList;
