+++ Development/src/pages/receivers/ReceiversShow.js
@@ -1,36 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { Link, Route } from 'react-router-dom';
 import {
-    ArrayField,
-    BooleanField,
-    BooleanInput,
-    Button,
-    ChipField,
-    Edit,
-    FormDataConsumer,
-    FunctionField,
-    LongTextInput,
-    ReferenceField,
-    SaveButton,
-    SelectInput,
-    ShowButton,
-    ShowController,
-    ShowView,
-    SimpleForm,
-    SimpleShowLayout,
-    TextField,
-    TextInput,
-    Title,
-    Toolbar,
-    withDataProvider,
-} from 'react-admin';
-import get from 'lodash/get';
-import set from 'lodash/set';
-import Cookies from 'universal-cookie';
-import {
     AppBar,
-    Card,
-    CardContent,
     Tab,
     Table,
     TableBody,
@@ -40,165 +11,46 @@ import {
     TableRow,
     Tabs,
 } from '@material-ui/core';
+import {
+    ArrayField,
+    BooleanField,
+    Button,
+    ChipField,
+    FunctionField,
+    ReferenceField,
+    ShowButton,
+    ShowController,
+    ShowView,
+    SimpleShowLayout,
+    TextField,
+    withDataProvider,
+} from 'react-admin';
+import get from 'lodash/get';
 import CheckIcon from '@material-ui/icons/Check';
 import ClearIcon from '@material-ui/icons/Clear';
-import dataProvider from '../dataProvider';
-import PaginationButton from '../components/PaginationButton';
-import FilterField from '../components/FilterField';
-import TAIField from '../components/TAIField';
-import MapTags from '../components/TagsField';
-import JsonIcon from '../components/JsonIcon';
-import ReceiverTransportParamsCardsGrid from '../components/ReceiverTransportParams';
-import ConnectionShowActions from '../components/ConnectionShowActions';
-import ConnectionEditActions from '../components/ConnectionEditActions';
-import JSONViewer from '../components/JSONViewer';
-import TransportFileViewer from '../components/TransportFileViewer';
-import ItemArrayField from '../components/ItemArrayField';
-import MakeConnection from '../components/MakeConnection';
+import ConnectionShowActions from '../../components/ConnectionShowActions';
+import FilterField from '../../components/FilterField';
+import ItemArrayField from '../../components/ItemArrayField';
+import JSONViewer from '../../components/JSONViewer';
+import MakeConnection from '../../components/MakeConnection';
+import PaginationButton from '../../components/PaginationButton';
+import QueryVersion from '../../components/QueryVersion';
+import ReceiverTransportParamsCardsGrid from './ReceiverTransportParams';
+import MapTags from '../../components/TagsField';
+import TAIField from '../../components/TAIField';
+import TransportFileViewer from '../../components/TransportFileViewer';
+import dataProvider from '../../dataProvider';
 
-const cookies = new Cookies();
-
-export const ReceiversList = () => {
-    const firstLoad = async () => {
-        const params = {
-            filter: {},
-        };
-        const dataObject = await dataProvider('GET_LIST', 'receivers', params);
-        setData(dataObject);
-    };
-
-    const [data, setData] = useState(firstLoad);
-
-    const nextPage = async label => {
-        const dataObject = await dataProvider(label, 'receivers');
-        setData(dataObject);
-    };
-
-    const [filterState, setFilterState] = useState({});
-
-    const changeFilter = async (filterValue, name) => {
-        let filter = filterState;
-        if (filterValue) {
-            filter[name] = filterValue;
-        } else {
-            delete filter[name];
-        }
-        const filteredDataObject = await dataProvider('GET_LIST', 'receivers', {
-            filter: filter,
-        });
-        setFilterState(filter);
-        setData(filteredDataObject);
-    };
-
-    const clearFilter = async () => {
-        setData(firstLoad());
-        setFilterState({});
-    };
-
-    if (data.hasOwnProperty('data')) {
-        return (
-            <Card>
-                <Title title={'Receivers'} />
-                <CardContent>
-                    <Button
-                        label={'Raw'}
-                        href={data.url}
-                        style={{ float: 'right' }}
-                        title={'View raw'}
-                    >
-                        <JsonIcon />
-                    </Button>
-                    <Table>
-                        <TableHead>
-                            <TableRow>
-                                <TableCell
-                                    style={{
-                                        minWidth: '240px',
-                                        paddingLeft: '32px',
-                                    }}
-                                >
-                                    Label{' '}
-                                    <FilterField
-                                        name="label"
-                                        setFilter={changeFilter}
-                                    />
-                                </TableCell>
-                                <TableCell style={{ minWidth: '245px' }}>
-                                    Format{' '}
-                                    <FilterField
-                                        name="format"
-                                        setFilter={changeFilter}
-                                    />
-                                </TableCell>
-                                <TableCell style={{ minWidth: '260px' }}>
-                                    Transport{' '}
-                                    <FilterField
-                                        name="transport"
-                                        setFilter={changeFilter}
-                                    />
-                                </TableCell>
-                                {QueryVersion() >= 'v1.2' && (
-                                    <TableCell>Active</TableCell>
-                                )}
-                            </TableRow>
-                        </TableHead>
-                        <TableBody>
-                            {data.data.map(item => (
-                                <TableRow key={item.id}>
-                                    <TableCell component="th" scope="row">
-                                        <ShowButton
-                                            style={{
-                                                textTransform: 'none',
-                                            }}
-                                            basePath="/receivers"
-                                            record={item}
-                                            label={item.label}
-                                        />
-                                    </TableCell>
-                                    <TableCell>{item.format}</TableCell>
-                                    <TableCell>{item.transport}</TableCell>
-                                    {QueryVersion() >= 'v1.2' && (
-                                        <TableCell>
-                                            {item.subscription.active ? (
-                                                <CheckIcon />
-                                            ) : (
-                                                <ClearIcon />
-                                            )}
-                                        </TableCell>
-                                    )}
-                                </TableRow>
-                            ))}
-                        </TableBody>
-                    </Table>
-                    <br />
-                    <PaginationButton label="FIRST" nextPage={nextPage} />
-                    <PaginationButton label="PREV" nextPage={nextPage} />
-                    <PaginationButton label="NEXT" nextPage={nextPage} />
-                    <PaginationButton label="LAST" nextPage={nextPage} />
-                    <Button
-                        onClick={() => clearFilter()}
-                        label="Clear All Filters"
-                    />
-                </CardContent>
-            </Card>
-        );
-    } else {
-        return <div />;
-    }
-};
-
-const ReceiversTitle = ({ record }) => {
-    return (
-        <span>
-            Receiver:{' '}
-            {record
-                ? record.label
-                    ? `${record.label}`
-                    : `${record.id}`
-                : 'Unknown'}
-        </span>
-    );
-};
+const ReceiversTitle = ({ record }) => (
+    <span>
+        Receiver:{' '}
+        {record
+            ? record.label
+                ? `${record.label}`
+                : `${record.id}`
+            : 'Unknown'}
+    </span>
+);
 
 const ChipConditionalLabel = ({ record, source, ...props }) => {
     props.clickable = true;
@@ -211,12 +63,7 @@ const ChipConditionalLabel = ({ record, source, ...props }) => {
     ) : null;
 };
 
-const QueryVersion = () => {
-    let url = cookies.get('Query API');
-    return url.match(/([^/]+)(?=\/?$)/g)[0];
-};
-
-const ReceiversShowComponent = props => {
+const ReceiversShow = props => {
     return (
         <ShowController {...props}>
             {controllerProps => (
@@ -483,127 +330,6 @@ const ShowStagedTab = ({ controllerProps, ...props }) => {
     );
 };
 
-const PostEditToolbar = props => (
-    <Toolbar {...props}>
-        <SaveButton label="Stage" />
-    </Toolbar>
-);
-
-export const ReceiversEdit = props => {
-    return (
-        <div>
-            <AppBar position="static" color="default">
-                <Tabs
-                    value={props.location.pathname}
-                    indicatorColor="primary"
-                    textColor="primary"
-                >
-                    <Tab
-                        label="Summary"
-                        component={Link}
-                        to={`${props.basePath}/${props.id}/show/`}
-                    />
-                    <Tab
-                        label="Active"
-                        component={Link}
-                        to={`${props.basePath}/${props.id}/show/active`}
-                    />
-                    <Tab
-                        label="Staged"
-                        value={`${props.match.url}`}
-                        component={Link}
-                        to={`${props.basePath}/${props.id}/show/staged`}
-                    />
-                    <Tab
-                        label="Connect"
-                        component={Link}
-                        to={`${props.basePath}/${props.id}/show/connect`}
-                    />
-                </Tabs>
-            </AppBar>
-            <Route
-                exact
-                path={`${props.basePath}/${props.id}/`}
-                render={() => <EditStagedTab {...props} />}
-            />
-        </div>
-    );
-};
-
-const EditStagedTab = props => (
-    <Edit
-        {...props}
-        undoable={false}
-        title={<ReceiversTitle />}
-        actions={<ConnectionEditActions id={props.id} />}
-    >
-        <SimpleForm
-            toolbar={<PostEditToolbar />}
-            redirect={`/receivers/${props.id}/show/staged`}
-        >
-            <TextInput label="Sender ID" source="$staged.sender_id" />
-            <BooleanInput
-                label="Master Enable"
-                source="$staged.master_enable"
-            />
-            <SelectInput
-                label="Activation Mode"
-                source="$staged.activation.mode"
-                choices={[
-                    { id: null, name: <ClearIcon /> },
-                    {
-                        id: 'activate_immediate',
-                        name: 'activate_immediate',
-                    },
-                    {
-                        id: 'activate_scheduled_relative',
-                        name: 'activate_scheduled_relative',
-                    },
-                    {
-                        id: 'activate_scheduled_absolute',
-                        name: 'activate_scheduled_absolute',
-                    },
-                ]}
-            />
-            <FormDataConsumer>
-                {({ formData, ...rest }) => {
-                    switch (get(formData, '$staged.activation.mode')) {
-                        case 'activate_scheduled_relative':
-                            return (
-                                <TextInput
-                                    label="Requested Time"
-                                    source="$staged.activation.requested_time"
-                                    {...rest}
-                                />
-                            );
-                        case 'activate_scheduled_absolute':
-                            return (
-                                <TextInput
-                                    label="Requested Time"
-                                    source="$staged.activation.requested_time"
-                                    {...rest}
-                                />
-                            );
-                        default:
-                            set(
-                                formData,
-                                '$staged.activation.requested_time',
-                                null
-                            );
-                            return null;
-                    }
-                }}
-            </FormDataConsumer>
-            <ReceiverTransportParamsCardsGrid />
-            <LongTextInput
-                label="Transport File"
-                source="$staged.transport_file.data"
-                resettable
-            />
-        </SimpleForm>
-    </Edit>
-);
-
 const ConnectionManagementTab = ({
     controllerProps,
     receiverData,
@@ -682,7 +408,7 @@ const ConnectionManagementTab = ({
                             <TableCell>Flow</TableCell>
                             <TableCell>Device</TableCell>
                             {QueryVersion() >= 'v1.2' && (
-                                <TableCell>Subscription Active</TableCell>
+                                <TableCell>Active</TableCell>
                             )}
                             <TableCell>Connect</TableCell>
                         </TableRow>
@@ -796,5 +522,4 @@ const ConnectionManagementTab = ({
     );
 };
 
-const ReceiversShow = withDataProvider(ReceiversShowComponent);
-export { ReceiversShow };
+export default withDataProvider(ReceiversShow);
