+++ Development/src/pages/receivers/index.js
@@ -1,0 +1,5 @@
+import ReceiversEdit from './ReceiversEdit';
+import ReceiversList from './ReceiversList';
+import ReceiversShow from './ReceiversShow';
+
+export { ReceiversEdit, ReceiversList, ReceiversShow };
