+++ Development/src/pages/senders/index.js
@@ -1,0 +1,5 @@
+import SendersEdit from './SendersEdit';
+import SendersList from './SendersList';
+import SendersShow from './SendersShow';
+
+export { SendersEdit, SendersList, SendersShow };
