+++ Development/src/pages/sources/index.js
@@ -1,0 +1,4 @@
+import SourcesList from './SourcesList';
+import SourcesShow from './SourcesShow';
+
+export { SourcesList, SourcesShow };
