+++ Development/src/pages/subscriptions/SubscriptionsList.js
@@ -1,20 +1,6 @@
 import React, { useState } from 'react';
 import {
-    BooleanField,
-    Button,
-    FunctionField,
-    ListButton,
-    Show,
-    ShowButton,
-    SimpleShowLayout,
-    TextField,
-    Title,
-} from 'react-admin';
-import { hr } from '@material-ui/core';
-import Cookies from 'universal-cookie';
-import {
     Card,
-    CardActions,
     CardContent,
     Table,
     TableBody,
@@ -22,16 +8,13 @@ import {
     TableHead,
     TableRow,
 } from '@material-ui/core';
-import dataProvider from '../dataProvider';
-import PaginationButton from '../components/PaginationButton';
-import FilterField from '../components/FilterField';
-import UrlField from '../components/URLField';
-import MapTags from '../components/TagsField';
-import JsonIcon from '../components/JsonIcon';
+import { BooleanField, Button, ShowButton, Title } from 'react-admin';
+import FilterField from '../../components/FilterField';
+import JsonIcon from '../../components/JsonIcon';
+import PaginationButton from '../../components/PaginationButton';
+import dataProvider from '../../dataProvider';
 
-const cookies = new Cookies();
-
-export const SubscriptionsList = () => {
+const SubscriptionsList = () => {
     const firstLoad = async () => {
         const params = {
             filter: {},
@@ -170,64 +153,4 @@ export const SubscriptionsList = () => {
     }
 };
 
-const SubscriptionsTitle = ({ record }) => {
-    return (
-        <span>
-            Subscription:{' '}
-            {record
-                ? record.label
-                    ? `${record.label}`
-                    : `${record.id}`
-                : 'Unknown'}
-        </span>
-    );
-};
-
-const cardActionStyle = {
-    zIndex: 2,
-    float: 'right',
-};
-
-const SubscriptionsShowActions = ({ basePath, data, resource }) => (
-    <CardActions title={<SubscriptionsTitle />} style={cardActionStyle}>
-        {data ? (
-            <Button
-                label={'Raw'}
-                href={cookies.get('Query API') + '/' + resource + '/' + data.id}
-                title={'View raw'}
-            >
-                <JsonIcon />
-            </Button>
-        ) : null}
-        <ListButton basePath={basePath} />
-    </CardActions>
-);
-
-export const SubscriptionsShow = props => (
-    <Show
-        title={<SubscriptionsTitle />}
-        actions={<SubscriptionsShowActions />}
-        {...props}
-    >
-        <SimpleShowLayout>
-            <TextField source="id" label="ID" />
-            <TextField source="resource_path" />
-            <UrlField label="WebSocket Address" source="ws_href" />
-            <TextField
-                label="Max Update Rate (ms)"
-                source="max_update_rate_ms"
-            />
-            <FunctionField
-                label="Params"
-                render={record =>
-                    Object.keys(record.params).length > 0
-                        ? MapTags(record)
-                        : null
-                }
-            />
-            <hr />
-            <BooleanField source="persist" />
-            <BooleanField source="secure" />
-        </SimpleShowLayout>
-    </Show>
-);
+export default SubscriptionsList;
