+++ Development/src/pages/subscriptions/index.js
@@ -1,0 +1,4 @@
+import SubscriptionsList from './SubscriptionsList';
+import SubscriptionsShow from './SubscriptionsShow';
+
+export { SubscriptionsList, SubscriptionsShow };
